Added a new index file in the pages directory to export all pages for cleaner code
