--- shop-app/src/index.css
@@ -1,10 +1,0 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
-
-*{
-    box-sizing: border-box;
-    outline: none;
-    margin: 0;
-    padding: 0;
-}
