+++ shop-app/src/main.tsx
@@ -1,10 +1,9 @@
-import React from "react";
 import { Store } from "./redux/Store";
 import { Provider } from "react-redux";
 import { createRoot } from "react-dom/client";
-import App from "./App";
-import "./index.css";
-import { ReactQueryProvider } from "./provider/ReactQueryProvider";
+import {App} from "./pages";
+import "./styles/index.css";
+import { ReactQueryProvider } from "./providers/ReactQueryProvider";
 
 const rootElem = document.getElementById("root");
 if (rootElem) {
