+++ shop-app/src/pages/404/Page404.tsx
@@ -1,4 +1,3 @@
-import React from "react";
 const Page404 = () => {
     return (
         <>
