+++ shop-app/src/pages/handleCart/HandleCartPage.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
-import CartPage from './CartPage';
-import LoginPage from "./LoginPage";
+import {CartPage, LoginPage} from "../index";
 
 const HandleCartPage: React.FC = () => {
     const isLoggedIn: string = localStorage.getItem("isLoggedIn") || "";
