+++ shop-app/src/pages/index.tsx
@@ -1,0 +1,14 @@
+import App from "./app/App";
+import Page404 from "./404/Page404";
+import CartPage from "./cart/CartPage";
+import CheckoutPage from "./checkout/CheckoutPage";
+import EmptyCartPage from "./empty/EmptyCartPage";
+import HandleCartPage from "./handleCart/HandleCartPage";
+import HomePage from "./home/HomePage";
+import LoginPage from "./login/LoginPage";
+import ProductPage from "./product/ProductPage";
+import ProductsListPage from "./productList/ProductsListPage";
+
+
+
+export {App, Page404, CartPage, CheckoutPage, EmptyCartPage, HandleCartPage, HomePage, LoginPage, ProductPage, ProductsListPage};
