+++ shop-app/src/pages/productList/ProductsListPage.tsx
@@ -1,5 +1,4 @@
-import React from "react";
-import Products from "../components/Products";
+import {Products} from "../../components";
 
 function ProductsListPage() {
   return (
